Leveraging Machine Learning to Combat Insurance Fraud in the Auto Insurance Industry
The primary objective of this project is to develop a predictive model that can classify
insurance claims as either fraudulent or legitimate based on the provided data. This involves
creating a binary classification system where the model assigns a label of "fraudulent" or
"non-fraudulent" to each insurance claim. The scope of the problem includes analyzing
historical data of insurance policies, customer information, and accident details to train the
predictive model. The model will then be deployed to predict the likelihood of fraud for new
insurance claims as they are submitted.